Split ConfigurationProvider and ConfigurationPersister
