+++ src/Integration.Vsix/AsmRef_Integration.Vsix_Baseline_WithoutStrongNames.txt
@@ -1,7 +1,7 @@
 ---
 ################################
 # Assembly references report
-# Report date/time: 2020-10-27T14:54:08.8664584Z
+# Report date/time: 2020-10-29T09:42:23.0046223Z
 ################################
 #
 # Generated by Devtility CheckAsmRefs v0.11.0.223
@@ -19,7 +19,7 @@
 #
 ################################
 
-# Base directory: C:\Projects\SLVS\sonarlint-visualstudio\src\Integration.Vsix\bin\Debug\VS2019\net46
+# Base directory: C:\Projects\sonarlint-visualstudio\src\Integration.Vsix\bin\Debug\VS2019\net46
 Include patterns:
 - 'Sonar*'
 Exclude patterns:
